component-base/logs: refactor registry
Moving the registry into its own package will avoid dependency cycles in future
PRs.

Creating loggers through a factory instead of storing pre-created instances
will make it possible to create the loggers differently depending on
configuration parameters.

The factory can also be used to provide additional meta data before creating
instances.
